requirements.txt
google-adk
pydantic>=2.7
pytest>=8.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.13.0
google-generativeai>=0.3.0
mistralai>=0.1.0
cohere>=4.0.0
requests>=2.31.0